# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tracking daily activities, streaks, and analytics.\n\n## Features\n- User authentication with JWT tokens\n- Activity tracking with customizable types\n- Streak tracking to maintain consistency\n- Weekly analytics with daily breakdowns\n- Privacy controls for user profiles\n- Tile configuration for dashboard customization\n\n## Rate Limits\n- Login/Register: 5 requests per minute per IP\n- Password Reset: 3 requests per minute per IP\n- General API: 60 requests per minute per user",
        "title": "Growth Tracker API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/aman1117/growth-tracker"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/get-bio": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get current user's bio",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Get bio",
                "responses": {
                    "200": {
                        "description": "Bio content",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.BioResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/get-privacy": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get current account privacy setting",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Get privacy setting",
                "responses": {
                    "200": {
                        "description": "Privacy setting",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.PrivacyResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get full profile of authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "User profile",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/update-bio": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update user bio (max 150 characters)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Update bio",
                "parameters": [
                    {
                        "description": "Bio content",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.UpdateBioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bio updated",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.BioResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/update-privacy": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Set account privacy (public or private)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Update privacy setting",
                "parameters": [
                    {
                        "description": "Privacy setting",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.UpdatePrivacyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Privacy updated",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.PrivacyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search for users by username (excludes private accounts)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Search users",
                "parameters": [
                    {
                        "description": "Search query",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.SearchUsersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.DataResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.UserDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_aman1117_backend_internal_dto.BioResponse": {
            "description": "Bio retrieval result",
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string",
                    "example": "Software developer | Coffee enthusiast"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.DataResponse": {
            "description": "Generic response with data payload",
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.ErrorResponse": {
            "description": "Generic error response",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "error_code": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.PrivacyResponse": {
            "description": "Privacy setting result",
            "type": "object",
            "properties": {
                "is_private": {
                    "type": "boolean",
                    "example": false
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_aman1117_backend_internal_dto.ProfileResponse": {
            "description": "User profile information",
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string",
                    "example": "Software developer"
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "profile_pic": {
                    "type": "string",
                    "example": "https://storage.blob.core.windows.net/pics/1/abc.jpg"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.SearchUsersRequest": {
            "description": "Search for users by username",
            "type": "object",
            "properties": {
                "username": {
                    "type": "string",
                    "example": "john"
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.UpdateBioRequest": {
            "description": "Bio update request",
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string",
                    "example": "Software developer | Coffee enthusiast"
                }
            }
        },
        "github_com_aman1117_backend_internal_dto.UpdatePrivacyRequest": {
            "description": "Privacy setting update request",
            "type": "object",
            "properties": {
                "is_private": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.UserDTO": {
            "description": "User information for search results",
            "type": "object",
            "properties": {
                "bio": {
                    "description": "Only included for public profiles",
                    "type": "string",
                    "example": "Software developer"
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_private": {
                    "type": "boolean",
                    "example": false
                },
                "profile_pic": {
                    "type": "string",
                    "example": "https://storage.blob.core.windows.net/pics/1/abc.jpg"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}